io: Add device parameter for fido_dev_io_close_t
Previously, fido_dev_io_close_t wouldn't take a parameter according
to the header file. However, the function should take a pointer to
some descriptor that was previously returned by fido_dev_io_open_t,
so the device can be closed properly.
This is also the way the functions are currently called in our code.

Co-authored-by: frcroth <frcroth@gmail.com>
Co-authored-by: konradh <49533343+konradh@users.noreply.github.com>