Introduce _FIDO_INTERNAL macro to differentiate external/internal functions
The _FIDO_INTERNAL macro allows us to differentiate between functions that
we use internally in the library and functions which we export to users.
The library itself should be compiled with the _FIDO_INTERNAL macro turned on,
but when using a compiled version of the library and header files to link
other code with, the _FIDO_INTERNAL macro should be turned off.

Co-authored-by: frcroth <frcroth@gmail.com>
Co-authored-by: konradh <49533343+konradh@users.noreply.github.com>