NFC: Fix APDU length calculation
Originally, apdu_len was the length of the whole buffer (with header).
Now, it is only the len of the payload. Surely, there are commands
where we don't need a payload, so the < 2 check does not make sense anymore.

We also remove the lc{1,2,3} fields in the APDU header, since we don't
use them. We always use short APDUs to save on stack space.

We remove the +1 in the length calculation which regards the expected
returned length. Since we (and the original libfido2) never set this field,
we don't need to send it.

Co-authored-by: felix-gohla <felix.gohla@student.hpi.de>
Co-authored-by: QuentinKuth <quentin.kuth@student.hpi.uni-potsdam.de>