Make fido_cbor_info_t struct of fixed size
Originally, the info struct would contain arbitrary sized
arrays which would be allocated using malloc.
However, since we don't have malloc, we need to make sure
that we can safely allocate our structs on the stack.
Hence, we make the info struct smaller and most importantly
of fixed size. However, this surely gives up on flexibility,
especially when using 64bit numbers as bitfields, as we can
have only 64 array entries. Nevertheless, this seems to be
enough for now.