Correctly handle CTAP status code in authenticator response
The first byte in the authenticator's CTAP response represents the
status code of the operation. Previously, we would not check this
byte and attempt to parse it with the rest of the buffer as CBOR.
Now, we check whether the first byte is set to success and parse the
rest of the buffer as CBOR afterwards.

Co-authored-by: felix-gohla <felix.gohla@student.hpi.de>
Co-authored-by: frcroth <frcroth@users.noreply.github.com>