Adjust CMakeLists for ESP-IDF
The last commit didn't really include the libmicrofido2
correctly in the build process, such that it would be able
to use the ESP-IDF provided functions. This seems to be an
issue with the ExternalProject directive, which we didn't
get to work so that ESP-IDF specific header files would be
included when building the libmicrofido2 for the ESP32 example.

Hence, we now modify the CMakeLists of the libmicrofido2 itself
to include settings for when it is built for an ESP_PLATFORM.
We also symlink the library as a component of the ESP32 example
project which lets us getting rid of the ExternalProject directive.

Really not pretty, but for now it works at least...

Co-authored-by: Felix Gohla <felix.gohla@student.hpi.de>