Increase stack size of ESP32 app_main to avoid crash
Previously, the stack size of the app_main task was the
default value, which appeared to be too low, as we got
output lines which indicated that the stack got smashed
and the ESP always restarted. The new value should be more
than enough and we didn't notice it crashing since.