Add measurement programs for Ed25519, Inflate, SHA256/512
This adds the programs for measuring the time it takes to
complete Ed25519, Inflate, and SHA256/512. Currently, these
only really compile for the ATmega1284P, but the GPIO
functionality can be adjusted for other boards as well.
Moreover, this also adds some Python scripts to generate the
input data for the measurement programs such as encryption
keys.