Add option to overwrite RNG implementation
Previously, the RNG implementation was provided by us, with
no way to overwrite it. Similar to the cryptographic algorithms,
users can now overwrite the RNG implementation by setting
USE_SOFTWARE_RNG to false and setting the fido_get_random function
in random.h accordingly.
This is especially useful when using a board that provides a hardware
RNG.